requirements.txt
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4>=4.12.2
tqdm>=4.65.0
PyQt6>=6.4.0
dnspython==2.4.2
matplotlib>=3.7.1
numpy>=1.24.3
google-cloud-aiplatform>=1.25.0
google-cloud-storage>=2.10.0
google-cloud-vision>=3.4.4
pdf2image>=1.16.3
pillow>=10.0.0
protobuf>=4.23.4
seaborn>=0.12.0
